Move LifecycleOwner to Java to fix compatibility
