+++ android/src/main/java/com/nt4f04und/android_content_provider/AndroidContentProviderMessageCodec.java
